+++ sort_test.go
@@ -7,11 +7,11 @@ import (
 )
 
 func TestMergeSort(t *testing.T) {
-	array := []int{1, -1, 7, 3, 2, 7, 6, 8, 5}
-	assert.Equal(t, []int{-1, 1, 2, 3, 5, 6, 7, 7, 8}, mergeSort(array))
+	array := []int64{1, -1, 7, 3, 2, 7, 6, 8, 5}
+	assert.Equal(t, []int64{-1, 1, 2, 3, 5, 6, 7, 7, 8}, mergeSort(array))
 }
 
 func TestQuickSort(t *testing.T) {
-	array := []int{1, -1, 7, 3, 2, 7, 6, 8, 5}
-	assert.Equal(t, []int{-1, 1, 2, 3, 5, 6, 7, 7, 8}, quickSort(array))
+	array := []int64{1, -1, 7, 3, 2, 7, 6, 8, 5}
+	assert.Equal(t, []int64{-1, 1, 2, 3, 5, 6, 7, 7, 8}, quickSort(array))
 }
